requirement.txt
torch
transformers
fastapi
uvicorn
python-multipart
pillow
